backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.10
python-jose[cryptography]>=3.3.0
PyJWT>=2.0.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.0
langchain-core>=0.3.27
langchain-community>=0.3.0
pydantic>=2.0.0
pdfplumber>=0.11.0
chromadb>=0.4.0
scikit-learn>=1.5.0
pypdf>=5.0.0
langchain-text-splitters>=0.3.0
pdf2image>=1.17.0
pillow>=11.0.0
pytesseract>=0.3.13
google-cloud-vision>=3.8.0
